GitIgnore folder path update
Updated gitignore to ignore the library, logs, etc folders that are inside the Tetrino Folder (as they should be) rather than in the root folder. I opened the project in unity, so some other files may have been created. Note that there's a second, case specific ignore statement for the Logs folder at the bottom